# docs/03-architecture/openapi.yaml
openapi: 3.1.0
info:
  title: LocalOps Desk Local API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:4317
paths:
...
  /api/hosts/{id}:
    put:
      summary: Update a local host configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated host
    delete:
      summary: Delete a local host configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deletion result
...